added easy_mpl file
