+++ ai4water/utils/__init__.py
@@ -3,3 +3,5 @@ from ai4water.utils.utils import make_model
 from ai4water.utils.visualizations import PlotResults
 
 from ai4water.utils.taylor_diagram import taylor_plot
+
+from .easy_mpl import plot, regplot, imshow, bar_chart
