+++ ai4water/utils/visualizations.py
@@ -1,5 +1,4 @@
 import os
-import random
 from typing import Union, Callable
 
 import numpy as np
@@ -15,8 +14,9 @@ try:
 except ModuleNotFoundError:
     plotly = None
 
-from .plotting_tools import save_or_show
-from ai4water.utils.utils import init_subplots
+from .plotting_tools import save_or_show, to_1d_array
+from ai4water.utils.easy_mpl import init_subplots
+from .easy_mpl import regplot
 
 # TODO add Murphy's plot as shown in MLAir
 # prediction_distribution aka actual_plot of PDPbox
@@ -26,24 +26,6 @@ from ai4water.utils.utils import init_subplots
 # show availability plot of data
 
 #
-regplot_combs = [
-    ['cadetblue', 'slateblue', 'darkslateblue'],
-    ['cadetblue', 'mediumblue', 'mediumblue'],
-    ['cornflowerblue', 'dodgerblue', 'darkblue'],
-    ['cornflowerblue', 'dodgerblue', 'steelblue'],
-    ['cornflowerblue', 'mediumblue', 'dodgerblue'],
-    ['cornflowerblue', 'steelblue', 'mediumblue'],
-    ['darkslateblue', 'aliceblue', 'mediumblue'],
-    ['darkslateblue', 'blue', 'royalblue'],
-    ['darkslateblue', 'blueviolet', 'royalblue'],
-    ['darkslateblue', 'darkblue', 'midnightblue'],
-    ['darkslateblue', 'mediumblue', 'darkslateblue'],
-    ['darkslateblue', 'midnightblue', 'mediumblue'],
-    ['seagreen', 'darkslateblue', 'cadetblue'],
-    ['cadetblue', 'darkblue', 'midnightblue'],
-    ['cadetblue', 'deepskyblue', 'cadetblue']
-]
-
 
 class Plot(object):
 
@@ -370,147 +352,6 @@ class PlotResults(Plot):
         plot_precision_recall_curve(estimator, x, y.reshape(-1, ))
         self.save_or_show(save, fname="plot_precision_recall_curve", where="results")
         return
-
-
-def regplot(
-        x: Union[np.ndarray, pd.DataFrame, pd.Series, list],
-        y: Union[np.ndarray, pd.DataFrame, pd.Series, list],
-        title: str = None,
-        show: bool = False,
-        annotation_key: str = None,
-        annotation_val: float = None,
-        line_color=None,
-        marker_color=None,
-        fill_color=None,
-        marker_size: int = 20,
-        ci: Union[int, None] = 95,
-        figsize: tuple = None,
-        xlabel: str = 'Observed',
-        ylabel: str = 'Predicted'
-):
-    """
-    Regpression plot with regression line and confidence interval
-
-    Arguments:
-        x : array like, the 'x' value.
-        y : array like
-        ci : confidence interval. Set to None if not required.
-        show : whether to show the plot or not
-        annotation_key : The name of the value to annotate with.
-        annotation_val : The value to annotate with.
-        marker_size :
-        line_color :
-        marker_color:
-        fill_color : only relevent if ci is not None.
-        figsize : tuple
-        title : name to be used for title
-        xlabel :
-        ylabel :
-
-    Example:
-        >>> from ai4water.datasets import arg_beach
-        >>> from ai4water.utils.visualizations import regplot
-        >>> data = arg_beach()
-        >>> regplot(data['pcp3_mm'], data['pcp6_mm'], show=True)
-    """
-    x = to_1d_array(x)
-    y = to_1d_array(y)
-
-    mc, lc, fc = random.choice(regplot_combs)
-    _metric_names = {'r2': '$R^2$'}
-    plt.close('all')
-
-    _, axis = plt.subplots(figsize=figsize or (6, 5))
-
-    axis.scatter(x, y, c=marker_color or mc,
-                 s=marker_size)  # set style options
-
-    if annotation_key is not None:
-        assert annotation_val is not None
-
-        plt.annotate(f'{annotation_key}: {round(annotation_val, 3)}',
-                     xy=(0.3, 0.95),
-                     xycoords='axes fraction',
-                     horizontalalignment='right', verticalalignment='top', fontsize=16)
-    _regplot(x,
-             y,
-             ax=axis,
-             ci=ci,
-             line_color=line_color or lc,
-             fill_color=fill_color or fc)
-
-    plt.xlabel(xlabel, fontsize=14)
-    plt.ylabel(ylabel, fontsize=14)
-    plt.title(title, fontsize=26)
-
-    if show:
-        plt.show()
-
-    return axis
-
-
-def _ci(a, which=95, axis=None):
-    """Return a percentile range from an array of values."""
-    p = 50 - which / 2, 50 + which / 2
-    return np.nanpercentile(a, p, axis)
-
-
-def reg_func(_x, _y):
-    return np.linalg.pinv(_x).dot(_y)
-
-
-def bootdist(f, args, n_boot=1000, **func_kwargs):
-
-    n = len(args[0])
-    integers = np.random.randint
-    boot_dist = []
-    for i in range(int(n_boot)):
-        resampler = integers(0, n, n, dtype=np.intp)  # intp is indexing dtype
-        sample = [a.take(resampler, axis=0) for a in args]
-        boot_dist.append(f(*sample, **func_kwargs))
-
-    return np.array(boot_dist)
-
-
-def _regplot(x, y, ax, ci=None, line_color=None, fill_color=None):
-
-    grid = np.linspace(np.min(x), np.max(x), 100)
-    x = np.c_[np.ones(len(x)), x]
-    grid = np.c_[np.ones(len(grid)), grid]
-    yhat = grid.dot(reg_func(x, y))
-
-    ax.plot(grid[:, 1], yhat, color=line_color)
-
-    if ci:
-        boots = bootdist(reg_func, args=[x, y], n_boot=1000).T
-
-        yhat_boots = grid.dot(boots).T
-
-        err_bands = _ci(yhat_boots, ci, axis=0)
-
-        ax.fill_between(grid[:, 1], *err_bands,
-                        facecolor=fill_color,
-                        alpha=.15)
-    return ax
-
-
-def to_1d_array(array_like) -> np.ndarray:
-
-    if array_like.__class__.__name__ in ['list', 'tuple', 'Series']:
-        return np.array(array_like)
-
-    elif array_like.__class__.__name__ == 'ndarray':
-        if array_like.ndim == 1:
-            return array_like
-        else:
-            assert array_like.size == len(array_like), f'cannot convert multidim ' \
-                                                       f'array of shape {array_like.shape} to 1d'
-            return array_like.reshape(-1, )
-
-    elif array_like.__class__.__name__ == 'DataFrame' and array_like.ndim == 2:
-        return array_like.values.reshape(-1,)
-    else:
-        raise ValueError(f'cannot convert object array {array_like.__class__.__name__}  to 1d ')
 
 
 def linear_model(
