+++ tests/test_utils/test_util_plots.py
@@ -5,7 +5,6 @@ import numpy as np
 import matplotlib.pyplot as plt
 
 from ai4water.utils.visualizations import murphy_diagram, fdc_plot
-from ai4water.utils.utils import imshow
 
 
 class TestMurphyDiagrams(unittest.TestCase):
@@ -54,11 +53,6 @@ class TestFDC(unittest.TestCase):
         assert isinstance(ax, plt.Axes)
         return
 
-class TestImshow(unittest.TestCase):
-
-    def test_imshow(self):
-        imshow(np.random.random((10, 10)), colorbar=True)
-        return
 
 
 if __name__ == "__main__":
